% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBWWorkshop2024_module1_data.R
\docType{data}
\name{CCLE_metabolomics}
\alias{CCLE_metabolomics}
\title{A \code{data.frame} containing metabolite expression data for the Cancer Cell Line
Encyclopedia (CCLE) dataset.}
\format{
\code{data.frame} object
}
\usage{
data(CCLE_metabolomics)
}
\description{
This is a subset of the CCLE metabolomics dataset from Ghandi et al. 2019 that includes
measurements of metabolite expression levels for a random selection of cell lines.
}
\references{
Ghandi, M., Huang, F. W., Jané-Valbuena, J., Kryukov, G. V., Lo, C. C., McDonald, E. R., Barretina, J., Gelfand, E. T., Bielski, C. M., Li, H., Hu, K., Andreev-Drakhlin, A. Y., Kim, J., Hess, J. M., Haas, B. J., Aguet, F., Weir, B. A., Rothberg, M. V., Paolella, B. R., … Golub, T. R. (2019). Next-generation characterization of the Cancer Cell Line Encyclopedia. Nature, 569(7757), 503–508. https://doi.org/10.1038/s41586-019-1186-3
}
\keyword{datasets}
