% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBWWorkshop2024_module1_data.R
\docType{data}
\name{CCLE_model_list}
\alias{CCLE_model_list}
\title{A \code{data.frame} containing metadata for cell lines used in the Cancer Cell
Line Encyclopedia (CCLE) project.}
\format{
\code{data.frame} object
}
\usage{
data(CCLE_model_list)
}
\description{
This is an annotation file for the CCLE dataset from Ghandi et al. 2019 containing a
list of cell line models used, including metadata for each cell line.
}
\references{
Ghandi, M., Huang, F. W., Jané-Valbuena, J., Kryukov, G. V., Lo, C. C., McDonald, E. R., Barretina, J., Gelfand, E. T., Bielski, C. M., Li, H., Hu, K., Andreev-Drakhlin, A. Y., Kim, J., Hess, J. M., Haas, B. J., Aguet, F., Weir, B. A., Rothberg, M. V., Paolella, B. R., … Golub, T. R. (2019). Next-generation characterization of the Cancer Cell Line Encyclopedia. Nature, 569(7757), 503–508. https://doi.org/10.1038/s41586-019-1186-3
}
\keyword{datasets}
