% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBWWorkshop2024_dummy_data.R
\docType{data}
\name{dummy_pset}
\alias{dummy_pset}
\title{A \code{PharmacoSet} object containing randomly generated data for demonstration
purposes.}
\format{
\code{PharmacoSet} object
\describe{
\item{sample}{50 samples}
\item{treatment}{22 treatments}
\item{molecularProfiles}{A \code{MultiAssayExperiment} with 2
\code{SummarizedExperiments}, both are RNASeq, but
we have manually applied a batch effect to one of them}
\item{treatmentResponse}{A \code{TreatmentResponseExperiment} with 184 total
experiments. Only the \code{raw} assay is included.
}
}
}
\usage{
data(dummy_pset)

CBWWorkshop2024::dummy_pset
}
\description{
A \code{PharmacoSet} object containing randomly generated data for demonstration
purposes.
}
\keyword{datasets}
