% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBWWorkshop2024_module1_data.R
\docType{data}
\name{GDSC_methylation_model_list}
\alias{GDSC_methylation_model_list}
\title{A \code{data.frame} containing cell line Sentrix array metadata for the Genomics
of Drug Sensitivity in Cancer (GDSC) dataset.}
\format{
\code{data.frame} object
}
\usage{
data(GDSC_methylation_model_list)
}
\description{
This is an annotation file of the GDSC methylation dataset from Iorio et al. 2016
containing Sentrix array platform metadata for cell lines with methylation profiling.
}
\references{
\enumerate{
\item Yang, W., Soares, J., Greninger, P., Edelman, E. J., Lightfoot, H., Forbes, S., Bindal, N., Beare, D., Smith, J. A., Thompson, I. R., Ramaswamy, S., & Futreal, P. A. (2013). Genomics of Drug Sensitivity in Cancer (GDSC): a resource for therapeutic biomarker discovery in cancer cells. Nucleic Acids Research, 41(D1), D955–D961. https://doi.org/10.1093/nar/gks1111
\item Iorio, F., Knijnenburg, T. A., Vis, D. J., Bignell, G. R., Menden, M. P., Schubert, M., Aben, N., Gonçalves, E., Barthorpe, S., Lightfoot, H., Cokelaer, T., Greninger, P., van Dyk, E., Chang, H., de Silva, H., Heyn, H., Deng, X., Egan, R. K., Liu, Q., Mironenko, T., … Garnett, M. J. (2016). A Landscape of Pharmacogenomic Interactions in Cancer. Cell, 166(3), 740–754. https://doi.org/10.1016/j.cell.2016.06.017
\item Garnett, M. J., Edelman, E. J., Heidorn, S. J., Greenman, C. D., Dastur, A., Lau, K. W., Greninger, P., Thompson, I. R., Luo, X., Soares, J., Liu, Q., Iorio, F., Surdez, D., Chen, L., Milano, R. J., Bignell, G. R., Tam, A. T., Davies, H., Stevenson, J. A., … Benes, C. H. (2012). Systematic identification of genomic markers of drug sensitivity in cancer cells. Nature, 483(7391), 570–575. https://doi.org/10.1038/nature11005
}
}
\keyword{datasets}
