% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBWWorkshop2024_NCI-Almanac_data.R
\docType{data}
\name{NCI_ALMANAC_pset}
\alias{NCI_ALMANAC_pset}
\title{The final \code{PharmacoSet} object containing the multiple parts of the NCI-ALMANAC
dataset.}
\format{
\code{PharmacoSet} object
\describe{
\item{sample}{60 samples}
\item{treatment}{31 treatments}
\item{molecularProfiles}{A \code{MultiAssayExperiment} with 4
\code{RangedSummarizedExperiments}, 'rna', 'mirna', 'rnaseq.comp', and 'rnaseq.iso'}
\item{treatmentResponse}{A \code{TreatmentResponseExperiment} with 10080 total
experiments. Assays are: 'response', 'profiles' and 'assay_metadata'.
}
}
}
\usage{
data(NCI_ALMANAC_pset)

CBWWorkshop2024::NCI_ALMANAC_pset
}
\description{
The final \code{PharmacoSet} object containing the multiple parts of the NCI-ALMANAC
dataset.
}
\keyword{datasets}
