% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBWWorkshop2024_NCI-Almanac_data.R
\docType{data}
\name{NCI_ALMANAC_sample_metadata}
\alias{NCI_ALMANAC_sample_metadata}
\title{A \code{data.frame} containing cell-line metadata manually curated as part of the
\code{PharmacoSet} creation process for the NCI-ALMANAC.}
\format{
\code{data.frame} object
}
\usage{
data(NCI_ALMANAC_sample_metadata)

CBWWorkshop2024::NCI_ALMANAC_sample_metadata
}
\description{
Sample level metadata on cell-lines from the NCI-ALMANAC, standardized to
use identifers from Cellosaurus and included manually curated tissue
annotations from Oncotree as well as Cellosaurus disease type annotations.
}
\references{
\enumerate{
\item Bairoch A. he Cellosaurus, a cell line knowledge resource. J. Biomol. Tech. 29:25-38(2018) DOI: 10.7171/jbt.18-2902-002; PMID: 29805321
\item Holbeck, S. L., Camalier, R., Crowell, J. A., Govindharajulu, J. P., Hollingshead, M., Anderson, L. W., Polley, E., Rubinstein, L., Srivastava, A., Wilsker, D., Collins, J. M., & Doroshow, J. H. (2017). The National Cancer Institute ALMANAC: A Comprehensive Screening Resource for the Detection of Anticancer Drug Pairs with Enhanced Therapeutic Activity. Cancer Research, 77(13), 3564–3576. https://doi.org/10.1158/0008-5472.CAN-17-0489
}
}
\keyword{datasets}
