% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBWWorkshop2024_NCI-Almanac_data.R
\docType{data}
\name{NCI_ALMANAC_treatment_metadata}
\alias{NCI_ALMANAC_treatment_metadata}
\title{A \code{data.frame} containing drug metadata manually curated as part of the
\code{PharmacoSet} creation process for the NCI-ALMANAC.}
\format{
\code{data.frame} object
}
\usage{
data(NCI_ALMANAC_treatment_metadata)

CBWWorkshop2024::NCI_ALMANAC_treatment_metadata
}
\description{
Treatment metadata on the various drugs used in the NCI-ALMANAC with additional
annotations included standardized drug identifiers which match the nomenclature
of other \code{PharmacoSet} object as well as links structural idenfiers such
as SMILES and InchiKeys as well as links to PubChem via the cid.
}
\references{
Holbeck, S. L., Camalier, R., Crowell, J. A., Govindharajulu, J. P., Hollingshead, M., Anderson, L. W., Polley, E., Rubinstein, L., Srivastava, A., Wilsker, D., Collins, J. M., & Doroshow, J. H. (2017). The National Cancer Institute ALMANAC: A Comprehensive Screening Resource for the Detection of Anticancer Drug Pairs with Enhanced Therapeutic Activity. Cancer Research, 77(13), 3564–3576. https://doi.org/10.1158/0008-5472.CAN-17-0489
}
\keyword{datasets}
